annotate `Transaction` with `must_use`
It turns out it's easy to get confused and thing that you're running
queries inside a transaction when in fact the transaction object gets
dropped and will be implicitly rolled back when the connection
terminates.

This PR adds a `must_use` annotation as an aid to developers that they
need to do something with the return value of `start_transaction`.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>